Update store to have its components brought in from different files
