+++ webapp/src/store/index.ts
@@ -1,11 +1,25 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import { AccountService, Auth } from '@/services';
-import router from '../router';
+import getters from './getters';
+import mutations from './mutations';
+import actions from './actions';
 
 Vue.use(Vuex)
 
 type TransactionType = 'deposit' | 'donation';
+
+export interface Token {
+  _id: string;
+  createdAt: Date;
+  updatedAt: Date;
+  expiresAt: Date;
+  user: string;
+}
+
+export interface LoginResult {
+  user: User;
+  token: Token;
+}
 
 export interface User {
   _id: string;
@@ -58,15 +72,13 @@ export interface Invitation {
   timestamp: Date;
 }
 
-interface AppState {
+export interface AppState {
   user?: User;
   beneficiaries: Beneficiary[];
   middlemen: Middleman[];
   transactions: Transaction[];
   invitations: Invitation[];
   invitation?: Invitation;
-  signinPhone?: string;
-  signupPhone?: string;
 }
 
 const state: AppState = {
@@ -76,251 +88,13 @@ const state: AppState = {
   transactions: [],
   invitations: [],
   invitation: undefined,
-  signinPhone: '',
-  signupPhone: '',
 }
 
-export default new Vuex.Store({
+const store = new Vuex.Store({
   state,
-  mutations: {
-    addTransaction(state, trx) {
-      state.transactions.push(trx);
-    },
-    addInvitation(state, invt) {
-      state.invitations.push(invt);
-    },
-    addBeneficiary(state, bnf) {
-      state.beneficiaries.push(bnf);
-    },
-    updateBeneficiary(state, bnf) {
-      state.beneficiaries[state.beneficiaries.findIndex(beneficiary => beneficiary._id === bnf._id)] = bnf;
-    },
-    addMiddleman(state, mdm) {
-      state.middlemen.push(mdm);
-    },
-    updateMiddleman(state, mdm) {
-      state.middlemen[state.middlemen.findIndex(middleman => middleman._id === mdm._id)] = mdm;
-    },
-    setUser(state, user) {
-      state.user = user
-    },
-    setBeneficiaries(state, beneficiaries) {
-      state.beneficiaries = beneficiaries
-    },
-    setMiddlemen(state, middlemen) {
-      state.middlemen = middlemen
-    },
-    setTransactions(state, transactions) {
-      state.transactions = transactions
-    },
-    setInvitations(state, invitations) {
-      state.invitations = invitations
-    },
-    removeInvitation(state, phone) {
-      const index = state.invitations.findIndex(invt => invt.invitee === phone);
-      if (index > - 1) state.invitations.splice(index, 1); 
-    },
-    setInvitation(state, invitation) {
-      state.invitation = invitation
-    },
-    setSigninPhone(state, phone) {
-      state.signinPhone = phone;
-    },
-    setSignupPhone(state, phone) {
-      state.signupPhone = phone;
-    }
-  },
-  getters: {
-    amountDeposited: ({ transactions }) => {
-      return transactions.filter(t => t.type == 'deposit' && t.amount > 0)
-        .map(t => t.amount)
-        .reduce((a, b) => a + b, 0);
-    },
-    totalAmountDonated: ({ transactions }) => {
-      return transactions.filter(t => t.type == 'donation' && t.amount > 0)
-        .map(t => t.amount)
-        .reduce((a, b) => a + b, 0);
-    },
-    accountBalance: ({ transactions }) => {
-      return transactions.map(t => t.type === 'donation' ? -1 * t.amount : t.amount).reduce((a, b) => a + b, 0);
-    },
-    peopleDonatedTo: ({ transactions, user }) => {
-      const recipients = transactions.filter(t => t.type === 'donation' && (t.to.length && user && t.to !== user._id))
-        .map(t => t.to);
-      const uniqueRecipients = new Set(recipients);
-      return uniqueRecipients.size;
-    },
-    donations: ({ transactions }) => {
-      return transactions.filter(t => t.type == 'donation').reverse();
-    },
-    numberOfBeneficiariesOwed: ({ beneficiaries, user }) => {
-      return beneficiaries.filter(bnf => {
-        if (user) {
-          const nominatorIndex = bnf.nominatedBy.findIndex(nominator => nominator._id === user._id);
-          return bnf.owed[nominatorIndex];
-        }
-        return false;
-      }).length;
-    },
-    numberOfBeneficiariesNotOwed: ({ beneficiaries, user }) => {
-      return beneficiaries.filter(bnf => {
-        if (user) {
-          const nominatorIndex = bnf.nominatedBy.findIndex(nominator => nominator._id === user._id);
-          return !bnf.owed[nominatorIndex];
-        }
-        return false;
-      }).length;
-    },
-    totalAmountOwedToBeneficiaries: ({ beneficiaries, user }) => {
-      if (user) {
-        return beneficiaries.reduce((total, bnf) => { 
-          const nominatorIndex = bnf.nominatedBy.findIndex(nominator => nominator._id === user._id);
-          return total + bnf.owed[nominatorIndex] 
-        }, 0);
-      }
-    }
-  },
-  actions: {
-    async getBeneficiaries({ commit}, _id: string) {
-      const beneficiaries = await AccountService.getBeneficiaries(_id);
-      commit('setBeneficiaries', beneficiaries);
-    },
-    async getMiddlemen({ commit}, _id: string) {
-      const middlemen = await AccountService.getMiddlemen(_id);
-      commit('setMiddlemen', middlemen);
-    },
-    async getTransactions({ commit}, _id: string) {
-      const transactions = await AccountService.getTransactions(_id);
-      commit('setTransactions', transactions);
-    },
-    async getInvitations({ commit, state}) {
-      if (state.user) {
-        const invitations = await AccountService.getInvitations(state.user.phone);
-        console.log('Inside getInvitations: ', invitations);
-        commit('setInvitations', invitations);
-      }
-    },
-    async donate({ commit, state }, { amount }: { amount: number }) {
-      if (state.user) {
-        const trx = await AccountService.donate(state.user._id, amount);
-        const updatedUser = await AccountService.updateUser({
-          ...state.user,
-          accountBalance: state.user.accountBalance + amount
-        });
-        commit('addTransaction', trx);
-        commit('setUser', updatedUser);
-      }
-    },
-    async nominateBeneficiary({ commit, state }, { nominator, beneficiary }: { nominator: string; beneficiary: string }) {
-      if (state.user) {
-        const result = await AccountService.getBeneficiary(beneficiary, state.user._id);
-        let bnf;
-        if (result) {
-          bnf = await AccountService.updateBeneficiary({ 
-            _id: nominator,
-            associatedDonorId: state.user._id,
-            timestamp: new Date()
-          }, result);
-          commit('updateBeneficiary', bnf);
-        }
-        else {
-          bnf = await AccountService.nominateBeneficiary({
-            _id: nominator,
-            associatedDonorId: state.user._id,
-            timestamp: new Date()
-          }, beneficiary);
-          commit('addBeneficiary', bnf);
-        }
-        
-      }
-    },
-    async appointMiddleman({ commit, state }, { middleman }: { middleman: string }) {
-      if (state.user) {
-        const result = await AccountService.getMiddleman(middleman, state.user._id);
-        let mdm;
-        if (result) {
-          mdm = await AccountService.updateMiddleman({ 
-            _id: state.user._id,
-            associatedDonorId: state.user._id,
-            timestamp: new Date()
-          }, result);
-          commit('updateMiddleman', mdm);
-        }
-        else {
-          mdm = await AccountService.appointMiddleman({
-            _id: state.user._id,
-            associatedDonorId: state.user._id,
-            timestamp: new Date()
-          }, middleman);
-          commit('addMiddleman', mdm);
-        }
-        const invt = await AccountService.sendInvitation(state.user, middleman);
-        console.log('invt: ', invt);
-        commit('addInvitation', invt);
-        console.log('state.invitations: ', state.invitations);
-      }
-    },
-    async resendInvitation({ commit, state }, { middleman }: { middleman: string }) {
-      if (state.user) {
-        const index = state.invitations.findIndex(invt => invt.invitee === middleman);
-        console.log('index: ', index);
-        console.log('state.invitations: ', state.invitations);
-        let invt;
-        if (index > - 1) {
-          invt = await AccountService.deleteInvitation(state.user, state.invitations[index]._id);
-          console.log('deleted invt: ', invt);
+  mutations,
+  getters,
+  actions
+});
 
-          commit('removeInvitation', middleman);
-        }
-
-        invt = await AccountService.sendInvitation(state.user, middleman);
-        console.log('resent invt: ', invt);
-        commit('addInvitation', invt);
-      } 
-    },
-    async getInvitation({ commit }, { path }: { path: string }) {
-      console.log('path in getInvitation: ', path);
-      const invitation = await AccountService.getInvitation(path);
-      commit('setInvitation', invitation);
-    },
-    async acceptInvitation({ commit }, { _id }: { _id: string }) {
-      console.log('_id in deleteInvitation: ', _id);
-      await AccountService.acceptInvitation(_id);
-      commit('setInvitation', null);
-    },
-    async doesUserExist({ commit }, { phone }: { phone: string }) {
-      console.log('phone in doesUserExist: ', phone);
-      const user = await AccountService.getUser(phone);
-      if (user) {
-        commit('setSigninPhone', phone);
-        router.push({ name: 'sign-in', params: { phone } });
-      }
-      else {
-        commit('setSignupPhone', phone);
-      }
-    },
-    async createUser({ commit }, { phone, email, role }: { phone: string; email: string; role: string }) {
-      console.log('In createUser: ', phone, email, role);
-      const user = await AccountService.createUser(phone, email, role);
-      if (user) {
-        commit('setSigninPhone', phone);
-        router.push({ name: 'sign-in', params: { phone } });
-      }
-      else {
-        commit('setSignupPhone', phone);
-      }
-    },
-    async signUserIn({ commit }, { phone, password }: { phone: string; password: string }) {
-      console.log('In signUserIn: ', phone, password);
-      const user = await AccountService.login(phone, password);
-      if (user) {
-        commit('setUser', user);
-        const token = JSON.stringify({ phone, password });
-        Auth.setAccessToken(token);
-        if (router.currentRoute.name !== 'home') router.push({ name: 'home' });
-      }
-    },
-  },
-  modules: {
-  }
-})
+export default store;
